move legacy bookmarks helper function to helper.js
